1974: Add tests
